Limit SVGR webpack package version range
Using a wildcard version range for the dependency breaks the usage when a new major version is released. This has been the case recently with SVGR v4 which will break projects when a [custom template][svgr-template] is used. It is caused by the API change from a custom AST parser to the powerful Babel AST API.
Even when the user uses a `package-lock.json` / `yarn.lock` file it will break as soon as it gets deleted because the `@svgr/webpack` package will now be resolved to the latest version.

This commit sets the version to the major version 4 range which means it will resolve all version `>=4.0.0`. Note that this can be seen as a breaking change, but with the wildcard version range used before the package broke every time a new major version oof any dependency has been released.

[svgr-template]: https://github.com/smooth-code/svgr#use-a-specific-template