Add limit to url-loader
Echo Gatsby's file-size limit for inlining image files which is set to 10000 bytes / 10kb. Without this addition, any svgs that are imported via URLs as opposed to SVGR React components will be inlined into packages, regardless of file-size, for bigger SVGs or pages with many SVGs this can be a serious performance issue.